fixes, but not correct yet with rabbit, score, user and quiz. needs fixing. changed long to int, that was a mistake -needs fixing

+++ quizservice/src/main/java/org/quizapp/quizservice/dtos/QuizDTO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class QuizDTO {
-    private Long quizId;
+    private int quizId;
     private String title;
     private String chapter;
     private String subject;
