+++ scoreservice/src/main/java/org/quizapp/scoreservice/repository/ScoreRepository.java
@@ -3,5 +3,5 @@ package org.quizapp.scoreservice.repository;
 import org.quizapp.scoreservice.model.Score;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScoreRepository extends JpaRepository<Score, Long> {
+public interface ScoreRepository extends JpaRepository<Score, Integer> {
 }
